requirements.txt
streamlit
python-dotenv
pandas
colorama 
PyYAML
pymongo
openpyxl